make home-details previous/next buttons to replace url
This makes it replace the current URL with next URL, doesn't create a
new state in the browser rather replace the current state. So if the
user browse different location detail and click the browser back button
it goes back to the home page, not the previous location he browsed.